requirements.txt
fastapi
SQLAlchemy
strawberry-graphql
uvicorn
pymysql
slack_sdk
python-dotenv
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
oauthlib
python-multipart
pyodbc
